+++ firmware/esp32/gdoor/src/gdoor_tx.h
@@ -23,7 +23,7 @@ namespace GDOOR_TX { //Namespace as we can only use it once
     extern uint16_t tx_state;
     void send(uint8_t *words, uint16_t len);
     void send(String str);
-    void setup(int txpin, int txenpin);
+    void setup(uint8_t txpin, uint8_t txenpin);
 };
 
 #endif
